requirements.txt
pip==22.0.4
wheel==0.37.1
setuptools==62.1.0
numpy==1.22.3
peewee==3.14.10
JPype1==1.3.0